Fix problem with saving tokens when deployed
