+++ src/popup/App.vue
@@ -452,6 +452,7 @@ export default {
           console.log(e)
         }
         
+        
         this.$store.dispatch('getAllUserTokens')
       }
